OPT: restructured test directory in preparation of extra unit tests, added tests for encryption and decryption logic
