+++ test/core/test_core_init.py
